Make specialize_template complain if not all replacements could be applied

+++ common.py
@@ -616,14 +616,23 @@ def recreate_directory(directory):
 def specialize_template(template_filename, destination_filename, replacements):
     template_file = open(template_filename, 'rb')
     lines = []
+    replaced = set()
 
     for line in template_file.readlines():
         for key in replacements:
-            line = line.replace(key, replacements[key])
+            replaced_line = line.replace(key, replacements[key])
+
+            if replaced_line != line:
+                replaced.add(key)
+
+            line = replaced_line
 
         lines.append(line)
 
     template_file.close()
+
+    if replaced != set(replacements.keys()):
+        raise Exception('Not all replacements for {0} have been applied'.format(template_filename))
 
     destination_file = open(destination_filename, 'wb')
     destination_file.writelines(lines)
